requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
asn1crypto==1.5.1
boto3==1.42.80
botocore==1.42.80
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
Django==6.0.1
django-ninja==1.6.2
django-silk==5.4.3
et_xmlfile==2.0.0
filelock==3.25.2
gprof2dot==2025.4.14
gunicorn==23.0.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
platformdirs==4.9.4
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
sentry-sdk==2.50.0
six==1.17.0
snowflake==1.12.0
snowflake-connector-python==4.4.0
snowflake._legacy==1.0.2
snowflake.core==1.12.0
sortedcontainers==2.4.0
sqlparse==0.5.5
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
 